TranslatorPro/requirements.txt
# ============================================
# TranslatorPro V3 — Зависимости
# Версия: 3.0.1 (refactored, без .rpyc)
# ============================================

# Перевод (оффлайн)
argostranslate>=1.9.0
transformers>=4.35.0
torch>=2.0.0

# RPA распаковка
unrpa>=1.2.0

# GUI
# customtkinter>=5.2.0  # Опционально: альтернатива tkinter

# Утилиты
requests>=2.31.0
python-dotenv>=1.0.0

# Обработка изображений
Pillow>=10.0.0

# Шрифты и текст
fonttools>=4.47.0
pyphen>=0.14.0

# Логирование
colorama>=0.4.6

# Дополнительно
beautifulsoup4>=4.12.0
regex>=2023.10.3
